+++ src/Microsoft.Management.Configuration.Processor/Constants/DirectiveConstants.cs
@@ -1,0 +1,23 @@
+﻿// -----------------------------------------------------------------------------
+// <copyright file="DirectiveConstants.cs" company="Microsoft Corporation">
+//     Copyright (c) Microsoft Corporation. Licensed under the MIT License.
+// </copyright>
+// -----------------------------------------------------------------------------
+
+namespace Microsoft.Management.Configuration.Processor.Constants
+{
+    /// <summary>
+    /// Directives.
+    /// </summary>
+    internal static class DirectiveConstants
+    {
+#pragma warning disable SA1600 // ElementsMustBeDocumented
+        public const string MaxVersion = "maxVersion";
+        public const string MinVersion = "minVersion";
+        public const string Module = "module";
+        public const string ModuleGuid = "moduleGuid";
+        public const string Repository = "repository";
+        public const string Version = "version";
+#pragma warning restore SA1600 // ElementsMustBeDocumented
+    }
+}
