--- src/Microsoft.Management.Configuration.Processor/Internals/Constants/PowerShellConstants.cs
@@ -1,68 +1,0 @@
-﻿// -----------------------------------------------------------------------------
-// <copyright file="PowerShellConstants.cs" company="Microsoft Corporation">
-//     Copyright (c) Microsoft Corporation. Licensed under the MIT License.
-// </copyright>
-// -----------------------------------------------------------------------------
-
-namespace Microsoft.Management.Configuration.Processor.Internals.Constants
-{
-    /// <summary>
-    /// Constants related to PowerShell.
-    /// </summary>
-    internal static class PowerShellConstants
-    {
-#pragma warning disable SA1600 // ElementsMustBeDocumented
-        public const string Core = "Core";
-
-        internal static class Variables
-        {
-            public const string PSEdition = "PSEdition";
-            public const string PSModulePath = "env:PSModulePath";
-        }
-
-        internal static class Modules
-        {
-            public const string PSDesiredStateConfiguration = "PSDesiredStateConfiguration";
-            public const string PSDesiredStateConfigurationVersion = "2.0.5";
-        }
-
-        internal static class Commands
-        {
-            public const string FindDscResource = "Find-DscResource";
-            public const string GetAuthenticodeSignature = "Get-AuthenticodeSignature";
-            public const string GetChildItem = "Get-ChildItem";
-            public const string GetCommand = "Get-Command";
-            public const string GetDscResource = "Get-DscResource";
-            public const string InstallModule = "Install-Module";
-            public const string InvokeDscResource = "Invoke-DscResource";
-            public const string SaveModule = "Save-Module";
-        }
-
-        internal static class Parameters
-        {
-            public const string Force = "Force";
-            public const string Guid = "GUID";
-            public const string InputObject = "InputObject";
-            public const string MaximumVersion = "MaximumVersion";
-            public const string Method = "Method";
-            public const string MinimumVersion = "MinimumVersion";
-            public const string Module = "Module";
-            public const string ModuleName = "ModuleName";
-            public const string ModuleVersion = "ModuleVersion";
-            public const string Name = "Name";
-            public const string Path = "Path";
-            public const string Property = "Property";
-            public const string Recurse = "Recurse";
-            public const string Repository = "Repository";
-            public const string RequiredVersion = "RequiredVersion";
-        }
-
-        internal static class DscMethod
-        {
-            public const string Get = "Get";
-            public const string Set = "Set";
-            public const string Test = "Test";
-        }
-#pragma warning restore SA1600 // ElementsMustBeDocumented
-    }
-}
