--- src/Microsoft.Management.Configuration.Processor/Internals/Helpers/StringHelpers.cs
@@ -1,24 +1,0 @@
-﻿// -----------------------------------------------------------------------------
-// <copyright file="StringHelpers.cs" company="Microsoft Corporation">
-//     Copyright (c) Microsoft Corporation. Licensed under the MIT License.
-// </copyright>
-// -----------------------------------------------------------------------------
-
-namespace Microsoft.Management.Configuration.Processor.Internals.Helpers
-{
-    /// <summary>
-    /// String helpers.
-    /// </summary>
-    internal static class StringHelpers
-    {
-        /// <summary>
-        /// Normalize string.
-        /// </summary>
-        /// <param name="value">Value.</param>
-        /// <returns>Normalized string.</returns>
-        public static string Normalize(string value)
-        {
-            return value.ToLowerInvariant();
-        }
-    }
-}
